requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
Flask-Migrate>=4.0.7
PyMySQL>=1.1.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0